style: PEP8 trailing whitespace fix for all other Python files
All changes are from:

sed -i 's/[ \t]*$//' "$1"

on appropriate Python files found in the codebase outside of the
'cf' directory (which was dealt with in previous commit). Note
'3.0.X/_downloads/tutorial.py' files have not been touched as we
can consider them as belonging to the previous cf versions.