Logging: basic root project logger with configurable levels
Note: setup and configuration is identical to that in cfdm, except
cf-python uses its own LOG_LEVEL constant in the CONSTANTS
dictionary to manage the log level (not the one from cfdm).

Includes loosening of the cfdm version contraint from an exact to a
minimum version, as agreed previously (changed now for development).